refactor(apim_tier): move flatteners + state setter to data source
Aligns api_instance_sla_tier with the codebase convention used by ame,
amq, dlb, apim — flatteners and the shared d.Set helper live in the
data source file, expanders stay in the resource. Pure code move, no
behaviour change.
